add optinal git option

+++ setup.sh
@@ -1,7 +1,16 @@
 #! /bin/sh
 
-rm -rf ./.git
+read -p "Use git locale? [y/n] " lgit
 
-git init
-git checkout -b main
-echo git remote add origin git@github.com:Raisess/${PWD##*/}
+if [ $lgit == "y" ]; then
+  rm -rf ./.git
+  git init
+  git checkout -b main
+
+  remote_repo=${PWD##*/}
+  read -p "Use git remote? (repository: $remote_repo) [y/n] " rgit
+
+  if [ $rgit == "y" ]; then
+    git remote add origin git@github.com:Raisess/$remote_repo
+  fi
+fi
